Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    double f;
+    float f;
     scanf("%f",&f);
     printf("You entered: %.4f\n",f);
     return 0;
